+++ spec/unit/ogr/internal_helpers_spec.rb
@@ -2,13 +2,9 @@
 
 require 'ogr/internal_helpers'
 
-module Tester
-  include OGR::InternalHelpers
-end
-
 RSpec.describe OGR::InternalHelpers do
   subject(:tester) do
-    Tester
+    extend OGR::InternalHelpers
   end
 
   describe '._boolean_access_flag' do